+++ sdk/clientcore/http-okhttp3/src/main/java/io/clientcore/http/okhttp3/implementation/OkHttpInputStreamRequestBody.java
@@ -4,7 +4,7 @@
 package io.clientcore.http.okhttp3.implementation;
 
 import io.clientcore.core.instrumentation.logging.ClientLogger;
-import io.clientcore.core.util.binarydata.InputStreamBinaryData;
+import io.clientcore.core.models.binarydata.InputStreamBinaryData;
 import okhttp3.MediaType;
 import okio.BufferedSink;
 import okio.Okio;