+++ sdk/maps/azure-maps-weather/src/main/java/com/azure/maps/weather/models/MinuteForecastSummary.java
@@ -37,7 +37,8 @@ public final class MinuteForecastSummary implements JsonSerializable<MinuteForec
     private String longPhrase;
 
     /*
-     * Numeric value representing an image that displays the `iconPhrase`. Please refer to [Weather services in Azure Maps](/azure/azure-maps/weather-services-concepts#weather-icons) for details.
+     * Numeric value representing an image that displays the `iconPhrase`. Please refer to [Weather services in Azure
+     * Maps](/azure/azure-maps/weather-services-concepts#weather-icons) for details.
      */
     private IconCode iconCode;
 
@@ -106,7 +107,7 @@ public final class MinuteForecastSummary implements JsonSerializable<MinuteForec
         jsonWriter.writeStringField("shortPhrase", this.shortDescription);
         jsonWriter.writeStringField("briefPhrase", this.briefDescription);
         jsonWriter.writeStringField("longPhrase", this.longPhrase);
-        jsonWriter.writeNumberField("iconCode", this.iconCode == null ? null : this.iconCode.toInt());
+        jsonWriter.writeNumberField("iconCode", this.iconCode == null ? null : this.iconCode.getValue());
         return jsonWriter.writeEndObject();
     }
 
@@ -133,7 +134,7 @@ public final class MinuteForecastSummary implements JsonSerializable<MinuteForec
                 } else if ("longPhrase".equals(fieldName)) {
                     deserializedMinuteForecastSummary.longPhrase = reader.getString();
                 } else if ("iconCode".equals(fieldName)) {
-                    deserializedMinuteForecastSummary.iconCode = IconCode.fromInt(reader.getInt());
+                    deserializedMinuteForecastSummary.iconCode = IconCode.fromValue(reader.getInt());
                 } else {
                     reader.skipChildren();
                 }