+++ sdk/identity/azure-identity-broker/CHANGELOG.md
@@ -1,9 +1,27 @@
 # Release History
 
-## 1.1.10 (2025-02-07)
+## 1.2.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
 
 ### Other Changes
 
+## 1.1.11 (2025-02-07)
+
+### Other Changes
+
+#### Dependency Updates
+- Upgraded `azure-identity` from `1.15.1` to `1.15.2`
+
+## 1.1.10 (2025-02-07)
+
+### Other Changes
+- 
+- Upgraded `azure-identity` from `1.15.0` to `1.15.1`
 - Upgraded `msal4j` from `1.17.2` to `1.19.0`
 
 ## 1.1.9 (2025-01-10)